fix warnings

+++ agrolib/interpolation/interpolation.cpp
@@ -2696,7 +2696,7 @@ float getFirstIntervalHeightValue(std::vector <Crit3DInterpolationDataPoint> &my
 bool isVectorNodataOrZero(std::vector <double> myVector)
 {
     bool myFlag = false;
-    for (int i = 0; i < myVector.size(); i++)
+    for (size_t i = 0; i < myVector.size(); i++)
     {
         if (isEqual(myVector[i], NODATA) || isEqual(myVector[i], 0))    myFlag = true;
     }
